+++ src/components/Map/TopMenu/HamburgerMenu.tsx
@@ -1,4 +1,3 @@
-import ChevronRightIcon from '@material-ui/icons/ChevronRight';
 import BrightnessAutoIcon from '@material-ui/icons/BrightnessAuto';
 import Brightness4Icon from '@material-ui/icons/Brightness4';
 import BrightnessHighIcon from '@material-ui/icons/BrightnessHigh';
@@ -9,18 +8,14 @@ import HelpIcon from '@material-ui/icons/Help';
 import styled from 'styled-components';
 import GetAppIcon from '@material-ui/icons/GetApp';
 import Router from 'next/router';
+import IconButton from '@material-ui/core/IconButton';
+import MenuIcon from '@material-ui/icons/Menu';
 import { useBoolState } from '../../helpers';
 import { t } from '../../../services/intl';
 import { useFeatureContext } from '../../utils/FeatureContext';
 import { useMapStateContext } from '../../utils/MapStateContext';
 import { getIdEditorLink } from '../../../utils';
 import { useUserThemeContext } from '../../../helpers/theme';
-
-const StyledChevronRightIcon = styled(ChevronRightIcon)`
-  color: ${({ theme }) => theme.palette.tertiary.main};
-  margin: -2px 0px -2px -1px !important;
-  font-size: 15px !important;
-`;
 
 const PencilIcon = styled(CreateIcon)`
   color: ${({ theme }) => theme.palette.action.active};
@@ -157,35 +152,38 @@ const ThemeSelection = () => {
 // https://github.com/mui-org/material-ui/issues/22912
 // https://github.com/mui-org/material-ui/issues?q=is%3Aissue+is%3Aopen+menuitem+keyboard
 
-export const MoreMenu = () => {
+export const HamburgerMenu = () => {
   const anchorRef = React.useRef();
   const [opened, open, close] = useBoolState(false);
 
   return (
     <>
       <Menu
-        id="more-menu"
+        id="hamburger-menu"
         anchorEl={anchorRef.current}
+        anchorOrigin={{
+          vertical: 'bottom',
+          horizontal: 'right',
+        }}
+        getContentAnchorEl={null}
         open={opened}
         onClose={close}
       >
-        <EditLink closeMenu={close} />
-        <AboutLink closeMenu={close} />
-        <InstallLink closeMenu={close} />
         <ThemeSelection />
+        <EditLink closeMenu={close} />
+        <InstallLink closeMenu={close} />
+        <AboutLink closeMenu={close} />
       </Menu>
-      <button
-        type="button"
-        className="linkLikeButton"
-        aria-controls="more-menu"
-        aria-haspopup="true"
-        onClick={open}
+      <IconButton
         ref={anchorRef}
+        aria-controls="hamburger-menu"
+        aria-haspopup="true"
         title={t('map.more_button_title')}
+        color="secondary"
+        onClick={open}
       >
-        {t('map.more_button')}
-        <StyledChevronRightIcon />
-      </button>
+        <MenuIcon />
+      </IconButton>
     </>
   );
 };
