+++ src/components/Map/TopMenu/TopMenu.tsx
@@ -1,9 +1,9 @@
 import IconButton from '@material-ui/core/IconButton';
 import AccountCircleIcon from '@material-ui/icons/AccountCircle';
-import MenuIcon from '@material-ui/icons/Menu';
 import React from 'react';
 import styled from 'styled-components';
-import { isDesktop } from '../helpers';
+import { isDesktop } from '../../helpers';
+import { HamburgerMenu } from './HamburgerMenu';
 
 const Wrapper = styled.span`
   vertical-align: top;
@@ -18,18 +18,18 @@ const Wrapper = styled.span`
   button:last-child {
     margin-left: -10px;
   }
+
+  svg {
+    filter: drop-shadow(0 0 2px #ffffff);
+  }
 `;
 
-export const TopMenu = () => {
-  return (
-    <Wrapper>
-      <IconButton>
-        <AccountCircleIcon />
-      </IconButton>
+export const TopMenu = () => (
+  <Wrapper>
+    <IconButton color="secondary">
+      <AccountCircleIcon />
+    </IconButton>
 
-      <IconButton>
-        <MenuIcon />
-      </IconButton>
-    </Wrapper>
-  );
-};
+    <HamburgerMenu />
+  </Wrapper>
+);
